requirements.txt
cachetools==5.5.2
certifi==2025.7.9
charset-normalizer==3.4.2
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==6.2.1
httplib2==0.22.0
idna==3.10
joblib==1.5.1
numpy==2.3.1
oauth2client==4.1.3
oauthlib==3.3.1
pandas==2.3.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.16.0
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
urllib3==2.5.0
